feat: club/main -> club/detail 폴더명 변경, 팀 정보 조회 및 곡 투표 조회 추가
