+++ src/pages/club/detail/clubSlide/VoteSlide.module.css
