+++ src/pages/club/detail/clubSlide/modalContent/VoteModal.tsx
