+++ src/types/univ.ts
@@ -1,0 +1,4 @@
+export interface University {
+  id: string;
+  name: string;
+}
